clean up codes
